This is a experiment of sort optimization.

1. change type of `key` in struct `orderByRow` from `[]types.Datum` to `[]*types.Datum` to avoid memcpy, i.e the `runtime.duffcopy` call, this will take 22% of the total `sort.Sort()` execution time
2. abandon usage of `CompareDatum()`, write compare functions for every type in every order(asc and desc).

in my test case, sort by 4 columns which include 2 int columns and 2 float-point columns, and without any index:

|  num records | optimized   | origin      |
|:-------------|------------:|------------:|
| 10           | 2151ns      | 3392ns      |
| 100          | 18267ns     | 35159ns     |
| 1000         | 150240ns    | 308438ns    |
| 10000        | 1635864ns   | 3845623ns   |
| 100000       | 37374192ns  | 67464658ns  |

and `runtime.duffcopy` call is reduced to zero